+++ easydel/modules/aya_vision/modeling_aya_vision_flax.py
@@ -100,8 +100,8 @@ class AyaVisionMultiModalProjector(nn.Module):
 	def __init__(
 		self,
 		config: AyaVisionConfig,
-		dtype: jnp.dtype = jnp.float32,
-		param_dtype: jnp.dtype = jnp.float32,
+		dtype: jnp.dtype = jnp.bfloat16,
+		param_dtype: jnp.dtype = jnp.bfloat16,
 		precision: jax.lax.PrecisionLike = None,
 		*,
 		rngs: nn.Rngs,
@@ -227,8 +227,8 @@ class AyaVisionForConditionalGeneration(EasyDeLBaseModule):
 	def __init__(
 		self,
 		config: AyaVisionConfig,
-		dtype: jnp.dtype = jnp.float32,
-		param_dtype: jnp.dtype = jnp.float32,
+		dtype: jnp.dtype = jnp.bfloat16,
+		param_dtype: jnp.dtype = jnp.bfloat16,
 		precision: jax.lax.PrecisionLike = None,
 		*,
 		rngs: nn.Rngs,
